moved bitmap code to cc file. added cast to unsigned type
